fix(logs): address PR review comments on time filter
